requirements.txt
# Core Dependencies
python-telegram-bot>=13.7
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.28.0
jdatetime>=4.1.0
pytz>=2022.1

# Technical Analysis
mplfinance>=0.12.10b0
ta-lib>=0.4.24

# Web and Async
aiohttp>=3.8.0

# Image Processing
pillow>=9.0.0

# Date/Time
python-dateutil>=2.8.2
